requirements.txt
openai>=1.0.0
python-telegram-bot==20.3
